+++ Gruntfile.js
@@ -80,7 +80,7 @@ module.exports = function (grunt) {
 
         clean: {
             build: {
-                src: ['build']
+                src: ['build', 'bower_components']
             },
             temp: {
                 src: ['build/src']
@@ -106,7 +106,7 @@ module.exports = function (grunt) {
                 overwrite: true,
                 src: ['src/config.xml'],
                 replacements: [{
-                    from: /version=\"[0-9]+\.[0-9]+\.[0-9]+(-dev)?\"/g,
+                    from: /version=\"[0-9]+\.[0-9]+\.[0-9]+(([ab]|rc)?[0-9]+)?(-dev)?\"/g,
                     to: 'version="<%= pkg.version %>"'
                 }]
             }
@@ -154,10 +154,9 @@ module.exports = function (grunt) {
                     helpers: ['src/test/helpers/*.js'],
                     vendor: ['bower_components/jquery/dist/jquery.js',
                              'bower_components/adapter.js/src/adapter.js',
-                             'bower_components/kurento-utils/js/kurento-utils.js',
                              'bower_components/bootstrap/dist/js/bootstrap.js',
+                             'node_modules/jasmine-jquery/lib/jasmine-jquery.js',
                              'bower_components/mock-socket/dist/mock-socket.js',
-                             'bower_components/jsPlumb/dist/js/dom.jsPlumb-1.7.2.js',
                              'src/test/vendor/*.js']
                 }
             },
@@ -184,12 +183,12 @@ module.exports = function (grunt) {
 
     grunt.loadNpmTasks('grunt-bower-task');
     grunt.loadNpmTasks('grunt-contrib-compress');
+    grunt.loadNpmTasks('grunt-contrib-copy');
     grunt.loadNpmTasks('grunt-contrib-jasmine');
     grunt.loadNpmTasks('grunt-contrib-jshint');
     grunt.loadNpmTasks('grunt-contrib-clean');
-    grunt.loadNpmTasks('grunt-contrib-copy');
-    grunt.loadNpmTasks('grunt-strip-code');
     grunt.loadNpmTasks('grunt-jscs');
+    grunt.loadNpmTasks('grunt-strip-code');
     grunt.loadNpmTasks('grunt-text-replace');
 
     grunt.registerTask('zip', 'compress:widget');