+++ src/doc/changelog.md
@@ -1,3 +1,14 @@
+v 0.3.2
+=============
+
+### General
+
+- Moved to use a [new
+  implementation](https://github.com/wirecloud-fiware/kurento-example-services-scala)
+  of the service
+- Updated the default server URL using https
+- New widget name: kurento-crowddetector. The previous widget name was: crowddetector.
+
 v 0.3.1
 ==============